Finished: BasicCorpse class was renamed to CompoundCorpse
