+++ src/dte/amongus/corpses/compound/components/blocks/BlockChangeComponent.java
@@ -1,9 +1,9 @@
-package dte.amongus.corpses.basic.components.blocks;
+package dte.amongus.corpses.compound.components.blocks;
 
 import org.bukkit.Material;
 import org.bukkit.block.Block;
 
-import dte.amongus.corpses.basic.BasicCorpse;
+import dte.amongus.corpses.compound.CompoundCorpse;
 import dte.amongus.utils.blocks.transformers.BlockTransformer;
 import dte.amongus.utils.blocks.transformers.restorers.Restorer;
 
@@ -19,7 +19,7 @@ public class BlockChangeComponent extends AbstractBlockComponent
 	 * @param corpse The corpse that consists of this block component.
 	 * @param block The block this component operates on.
 	 */
-	public BlockChangeComponent(BasicCorpse corpse, Block block, Material newMaterial) 
+	public BlockChangeComponent(CompoundCorpse corpse, Block block, Material newMaterial) 
 	{
 		this(corpse, block, newMaterial, false);
 	}
@@ -31,7 +31,7 @@ public class BlockChangeComponent extends AbstractBlockComponent
 	 * @param block The block this component operates on.
 	 * @param applyPhysics Whether the surroundings blocks will be affected from the material change. 
 	 */
-	public BlockChangeComponent(BasicCorpse corpse, Block block, Material newMaterial, boolean applyPhysics) 
+	public BlockChangeComponent(CompoundCorpse corpse, Block block, Material newMaterial, boolean applyPhysics) 
 	{
 		this(corpse, block, new BlockTransformer().toMaterial(newMaterial, applyPhysics));
 	}
@@ -43,7 +43,7 @@ public class BlockChangeComponent extends AbstractBlockComponent
 	 * @param block The block this component operates on.
 	 * @param transformer The transformer which will change the block.
 	 */
-	public BlockChangeComponent(BasicCorpse corpse, Block block, BlockTransformer transformer) 
+	public BlockChangeComponent(CompoundCorpse corpse, Block block, BlockTransformer transformer) 
 	{
 		super(corpse, block);
