+++ src/dte/amongus/corpses/compound/components/concretes/DeathSignComponent.java
@@ -1,17 +1,17 @@
-package dte.amongus.corpses.basic.components.concretes;
+package dte.amongus.corpses.compound.components.concretes;
 
 import static dte.amongus.utils.blocks.transformers.BlockTransformer.toSign;
 
 import org.bukkit.Material;
 import org.bukkit.block.Block;
 
-import dte.amongus.corpses.basic.BasicCorpse;
-import dte.amongus.corpses.basic.components.blocks.BlockChangeComponent;
+import dte.amongus.corpses.compound.CompoundCorpse;
+import dte.amongus.corpses.compound.components.blocks.BlockChangeComponent;
 import dte.amongus.games.players.Crewmate;
 
 public class DeathSignComponent extends BlockChangeComponent
 {
-	public DeathSignComponent(BasicCorpse corpse, Block block, Material signMaterial)
+	public DeathSignComponent(CompoundCorpse corpse, Block block, Material signMaterial)
 	{
 		super(corpse, block, toSign(signMaterial, createLinesFor(corpse.whoDied())));
 	}
